move editors to their own directory (start decreasing component granularity)
