+++ src/editors/ValueEditors.tsx
@@ -12,10 +12,10 @@ import type {
   EnumDeclaration,
   Typ,
   ValueDef,
-} from './generated/test_case';
+} from '../generated/test_case';
 import { ArrayEditor } from './ArrayEditor';
-import { assertUnreachable } from './util';
-import { getDefaultValue } from './defaults';
+import { assertUnreachable } from '../util';
+import { getDefaultValue } from '../defaults';
 
 // Helper to create a RuntimeValue from a RuntimeValueRaw, preserving attrs
 export function createRuntimeValue(
@@ -553,7 +553,7 @@ function StructEditor(props: StructEditorProps): ReactElement {
 
   return (
     <div className="struct-editor struct-container">
-      <div className="struct-header">{structDeclaration.struct_name}</div>
+      <div className="struct-name">{structDeclaration.struct_name}</div>
       {Array.from(fields.entries()).map(([fieldName, fieldType]) => {
         return (
           <div className="struct-section">
